pyproject.toml
[project]
name = "agent-engine-cicd-base"
version = "0.6.0"
description = "Base repo for deploying ADK apps to Agent Engine"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    # Pin ADK to test releases before locking upgrades
    "google-adk==1.17.0",
    "google-auth>=2.40.3,<3.0.0",
    "google-cloud-aiplatform[agent-engines]>=1.112.0,<2.0.0",
    "google-cloud-logging>=3.12.1,<4.0.0",
    # Limit google-cloud-storage to <3.0.0 due to conflicts in 3.x
    # NOTE: ADK 1.18.0 requires google-cloud-storage >=3.0.0
    "google-cloud-storage>=2.18.0,<3.0.0",
    "httpx>=0.28.1,<1.0.0",
    "opentelemetry-exporter-gcp-logging>=1.9.0a0,<2.0.0",
    # Pin the OTLP exporter for stability with the latest ADK telemetry features
    "opentelemetry-exporter-otlp-proto-grpc==1.37.0",
    "opentelemetry-instrumentation-google-genai>=0.4b0,<0.5",
    "opentelemetry-instrumentation-logging>=0.58b0,<0.59",
    "pydantic>=2.11.9,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "uvicorn>=0.35.0,<1.0.0",
]

[project.scripts]
deploy = "agent_engine_cicd_base.deployment.deploy_agent:deploy"
delete = "agent_engine_cicd_base.deployment.deploy_agent:delete"
register = "agent_engine_cicd_base.deployment.register_agent:main_register"
unregister = "agent_engine_cicd_base.deployment.register_agent:main_unregister"
list-agents = "agent_engine_cicd_base.deployment.register_agent:main_list"
local-agent = "agent_engine_cicd_base.server:main"
server = "agent_engine_cicd_base.server:main"
remote-agent = "agent_engine_cicd_base.deployment.run_remote_agent:main"
test-deploy = "agent_engine_cicd_base.deployment.deploy_agent:test_environment"
test-register = "agent_engine_cicd_base.deployment.register_agent:test_environment"
init-template = "agent_engine_cicd_base.deployment.init_template:main"

[dependency-groups]
dev = [
    "ruff>=0.12.7,<1.0.0",
    "mypy>=1.17.1,<2.0.0",
    "ipykernel>=6.30.1,<7.0.0",
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-mock>=3.0.0,<4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_engine_cicd_base"]
exclude = ["src/agent_engine_cicd_base/deployment"]

[tool.ruff]
src = ["src"]
target-version = "py313"
line-length = 88
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # Allow assert statements in tests

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
packages = ["agent_engine_cicd_base"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

# Deployment modules use evolving Google Cloud APIs with version compatibility issues
[[tool.mypy.overrides]]
module = "agent_engine_cicd_base.deployment.*"
disable_error_code = ["attr-defined", "arg-type", "import-untyped"]

# OpenTelemetry modules have limited type stubs
[[tool.mypy.overrides]]
module = "agent_engine_cicd_base.utils.observability"
disable_error_code = ["import-untyped"]

[tool.coverage.run]
branch = true
include = [
    "src/agent_engine_cicd_base/deployment/config.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
skip_empty = true
fail_under = 100
